+++ src/main/java/br/edu/ifgoiano/Empreventos/dto/RatingDTO.java
@@ -7,10 +7,10 @@ import jakarta.validation.constraints.Size;
 
 public class RatingDTO {
 
-    private Integer id;
+    private Long id;
 
     @NotNull(message = "O ID da inscrição é obrigatório.")
-    private Integer subscriptionId;
+    private Long subscriptionId;
 
     @NotNull(message = "A nota é obrigatória.")
     @Min(value = 1, message = "A nota mínima é 1.")
@@ -22,11 +22,11 @@ public class RatingDTO {
 
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -46,11 +46,11 @@ public class RatingDTO {
         this.comment = comment;
     }
 
-    public Integer getSubscriptionId() {
+    public Long getSubscriptionId() {
         return subscriptionId;
     }
 
-    public void setSubscriptionId(Integer subscriptionId) {
+    public void setSubscriptionId(Long subscriptionId) {
         this.subscriptionId = subscriptionId;
     }
 }
