+++ src/main/java/br/edu/ifgoiano/Empreventos/dto/SubscriptionResponseDTO.java
@@ -4,21 +4,21 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class SubscriptionResponseDTO {
-    private Integer id;
+    private Long id;
     private String status;
     private BigDecimal amountPaid;
     private LocalDateTime createdAt;
-    private Integer eventId;
+    private Long eventId;
     private String eventTitle;
-    private Integer listenerId;
+    private Long listenerId;
     private String listenerName;
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -46,11 +46,11 @@ public class SubscriptionResponseDTO {
         this.createdAt = createdAt;
     }
 
-    public Integer getEventId() {
+    public Long getEventId() {
         return eventId;
     }
 
-    public void setEventId(Integer eventId) {
+    public void setEventId(Long eventId) {
         this.eventId = eventId;
     }
 
@@ -62,11 +62,11 @@ public class SubscriptionResponseDTO {
         this.eventTitle = eventTitle;
     }
 
-    public Integer getListenerId() {
+    public Long getListenerId() {
         return listenerId;
     }
 
-    public void setListenerId(Integer listenerId) {
+    public void setListenerId(Long listenerId) {
         this.listenerId = listenerId;
     }
