+++ src/main/java/br/edu/ifgoiano/Empreventos/model/User.java
@@ -15,7 +15,7 @@ public class User implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(nullable = false, length = 100)
     private String name;
     @Column(nullable = false, length = 255)
@@ -94,7 +94,7 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(Integer id, String name, String password, String cpf_cnpj, String email, String phone_number, Boolean active, String avatar_url, String bio, Date created_at, Date updated_at, Date deleted_at, Set<UserRole> userRoles) {
+    public User(Long id, String name, String password, String cpf_cnpj, String email, String phone_number, Boolean active, String avatar_url, String bio, Date created_at, Date updated_at, Date deleted_at, Set<UserRole> userRoles) {
         this.id = id;
         this.name = name;
         this.password = password;
@@ -110,11 +110,11 @@ public class User implements Serializable {
         this.userRoles = userRoles;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
