+++ src/main/java/br/edu/ifgoiano/Empreventos/service/RatingService.java
@@ -22,7 +22,7 @@ public class RatingService {
    @Autowired
     private SubscriptionRepository subscriptionRepository;
 
-    public RatingDTO create(Integer subscriptionId, RatingDTO ratingDTO) {
+    public RatingDTO create(Long subscriptionId, RatingDTO ratingDTO) {
         var subscription = subscriptionRepository.findById(subscriptionId)
                 .orElseThrow(() -> new NoSuchElementException("Inscrição não encontrada com ID: " + subscriptionId));
         var rating = DataMapper.parseObject(ratingDTO, Rating.class);
@@ -31,12 +31,12 @@ public class RatingService {
         return DataMapper.parseObject(savedRating, RatingDTO.class);
     }
 
-    public List<RatingDTO> findBySubscriptionId(Integer subscriptionId) {
+    public List<RatingDTO> findBySubscriptionId(Long subscriptionId) {
         var ratings = ratingRepository.findBySubscriptionId(subscriptionId);
         return DataMapper.parseListObjects(ratings, RatingDTO.class);
     }
 
-    public void delete(Integer ratingId) {
+    public void delete(Long ratingId) {
         var rating = ratingRepository.findById(ratingId)
                 .orElseThrow(() -> new NoSuchElementException("Avaliação com ID " + ratingId + " não encontrada."));
         rating.setDeletedAt(LocalDateTime.now());
